refactor(route): extract route selection state hook
Extract route list/map selection state and synchronization effects into useRouteSelectionState.
Keep Route page focused on data composition and layout by moving sidebar, stop, vehicle, and tab interaction logic into the new hook.
